Reworked tutorials in ollama_demo_notebook_shared.ipynb
The download / setup portions are unchanged, but most of the tutorials have updated explanations and a fair amount of code changes. I felt that many of the descriptions were not only lacking, but actually would mislead the student as they built their "mental model" of what the code was doing.

I also gave up on getting tinyllama to give satisfactory responses; a quick switch to phi3.5 improved results dramatically. I ran on the "data100" Hub.